feat: refactor shared components, models, and services into separate ‘shared’ directory
