+++ src/app/admin-auth-guard.service.ts
@@ -1,5 +1,5 @@
-import { UserService } from './user.service';
-import { AuthService } from './auth.service';
+import { UserService } from 'shared/services/user.service';
+import { AuthService } from 'shared/services/auth.service';
 import { Injectable } from '@angular/core';
 import { CanActivate } from '@angular/router';
 import { map } from 'rxjs/operators';
