+++ src/app/admin/admin-orders/admin-orders.component.ts
@@ -1,5 +1,5 @@
-import { Order } from './../../models/order';
-import { OrderService } from './../../order.service';
+import { Order } from 'shared/models/order';
+import { OrderService } from 'shared/services/order.service';
 import { Component } from '@angular/core';
 
 @Component({
