+++ src/app/check-out/check-out.component.ts
@@ -1,5 +1,5 @@
-import { ShoppingCart } from './../models/shopping-cart';
-import { ShoppingCartService } from './../shopping-cart.service';
+import { ShoppingCart } from 'shared/models/shopping-cart';
+import { ShoppingCartService } from 'shared/services/shopping-cart.service';
 import { Component, OnInit } from '@angular/core';
 import { Observable } from 'rxjs';
