+++ src/app/login/login.component.ts
@@ -1,4 +1,4 @@
-import { AuthService } from './../auth.service';
+import { AuthService } from 'shared/services/auth.service';
 import { Component } from '@angular/core';
 
 @Component({
