+++ src/app/my-orders/my-orders.component.ts
@@ -1,7 +1,7 @@
 import { switchMap } from 'rxjs/operators';
-import { AuthService } from './../auth.service';
+import { AuthService } from 'shared/services/auth.service';
 import { Component, OnInit } from '@angular/core';
-import { OrderService } from '../order.service';
+import { OrderService } from 'shared/services/order.service';
 
 @Component({
   selector: 'app-my-orders',
