+++ src/app/shared/components/product-card/product-card.component.scss
