+++ src/app/shared/components/product-card/product-card.component.ts
@@ -1,7 +1,7 @@
-import { ShoppingCart } from '../models/shopping-cart';
-import { ShoppingCartService } from './../shopping-cart.service';
+import { ShoppingCart } from 'shared/models/shopping-cart';
+import { ShoppingCartService } from 'shared/services/shopping-cart.service';
 import { Component, Input } from '@angular/core';
-import { Product } from '../models/product';
+import { Product } from 'shared/models/product';
 
 @Component({
   selector: 'product-card',
