+++ src/app/shared/models/shopping-cart-item.ts
