+++ src/app/shared/models/shopping-cart.ts
