+++ src/app/shared/services/auth-guard.service.ts
@@ -1,5 +1,5 @@
 import { map } from 'rxjs/operators';
-import { AuthService } from './auth.service';
+import { AuthService } from 'shared/services/auth.service';
 import { Injectable } from '@angular/core';
 import { CanActivate, Router, RouterStateSnapshot } from '@angular/router';
 import { Observable } from 'rxjs';
