+++ src/app/shared/services/auth.service.ts
@@ -1,10 +1,10 @@
-import { UserService } from './user.service';
+import { UserService } from 'shared/services/user.service';
 import { Injectable } from '@angular/core';
 import { AngularFireAuth } from '@angular/fire/auth';
 import * as firebase from 'firebase';
 import { Observable, of } from 'rxjs';
 import { ActivatedRoute, Router } from '@angular/router';
-import { AppUser } from './models/app-user';
+import { AppUser } from 'shared/models/app-user';
 import { switchMap } from 'rxjs/operators';
 
 @Injectable({
