+++ src/app/shared/services/shopping-cart.service.ts
@@ -1,10 +1,10 @@
-import { documentToDomainObject } from './util/documentToDomainObject';
-import { ShoppingCartItem } from './models/shopping-cart-item';
+import { documentToDomainObject } from '../../util/documentToDomainObject';
+import { ShoppingCartItem } from 'shared/models/shopping-cart-item';
 import { take, map } from 'rxjs/operators';
 import { Injectable } from '@angular/core';
 import { AngularFireDatabase, AngularFireObject } from '@angular/fire/database';
-import { Product } from './models/product';
-import { ShoppingCart } from './models/shopping-cart';
+import { Product } from 'shared/models/product';
+import { ShoppingCart } from 'shared/models/shopping-cart';
 import { database } from 'firebase';
 import { Observable } from 'rxjs';
