+++ src/app/shipping-form/shipping-form.component.ts
@@ -1,16 +1,16 @@
-import { ShoppingCart } from './../models/shopping-cart';
+import { ShoppingCart } from 'shared/models/shopping-cart';
 import { Component, OnInit, OnDestroy, Input } from '@angular/core';
 import { Subscription } from 'rxjs';
 import { Router } from '@angular/router';
-import { AuthService } from '../auth.service';
-import { OrderService } from '../order.service';
+import { AuthService } from 'shared/services/auth.service';
+import { OrderService } from 'shared/services/order.service';
 import {
   FormBuilder,
   FormGroup,
   AbstractControl,
   Validators,
 } from '@angular/forms';
-import { Order } from '../models/order';
+import { Order } from 'shared/models/order';
 
 enum Label {
   NAME = 'Name',
