+++ src/app/shopping-cart/shopping-cart.component.ts
@@ -1,7 +1,7 @@
 import { Observable } from 'rxjs';
-import { ShoppingCartService } from './../shopping-cart.service';
+import { ShoppingCartService } from 'shared/services/shopping-cart.service';
 import { Component, OnInit } from '@angular/core';
-import { ShoppingCart } from '../models/shopping-cart';
+import { ShoppingCart } from 'shared/models/shopping-cart';
 
 @Component({
   selector: 'shopping-cart',
